cleanup tons of stuff
