+++ gatsby-config.js
@@ -18,7 +18,7 @@ module.exports = {
         icon: 'src/images/gatsby-icon.png', // This path is relative to the root of the site.
       },
     },
-    'gatsby-plugin-layout'
+    // 'gatsby-plugin-layout'
     // 'gatsby-plugin-offline',
   ],
 }
