+++ package.json
@@ -6,8 +6,8 @@
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
     "firebase": "^5.5.4",
-    "gatsby": "^2.0.19",
-    "gatsby-plugin-layout": "^1.0.3",
+    "gatsby": "^2.3.25",
+    "gatsby-plugin-layout": "^1.0.14",
     "gatsby-plugin-manifest": "^2.0.5",
     "gatsby-plugin-offline": "^2.0.5",
     "gatsby-plugin-react-helmet": "^3.0.0",
